requirements.txt
et_xmlfile==2.0.0
filelock==3.18.0
fsspec==2025.5.1
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
scipy
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
torchsummary
